01_agentic_chatbot_langchain/app/requirements.txt
numpy==1.26.4
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apprise==1.9.4
asgi-lifespan==2.1.0
asn1crypto==1.5.1
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
azure-common==1.1.28
azure-core==1.36.0
azure-identity==1.25.1
azure-mgmt-core==1.6.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==23.1.0
azure-mgmt-synapse==2.0.0
azure-storage-blob==12.27.0
azure-storage-file-datalake==12.22.0
bleach==6.2.0
blinker==1.9.0
boto3==1.40.59
botocore==1.40.59
cachetools==5.5.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
coolname==2.2.0
cryptography==45.0.7
cycler==0.12.1
databricks-sdk==0.65.0
dataclasses-json==0.6.7
dateparser==1.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
docker==7.1.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.12.0
fastapi==0.116.1
Flask==3.1.2
fonttools==4.59.2
frozenlist==1.6.2
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.2
griffe==1.14.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
humanize==4.13.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kiwisolver==1.4.9
langchain==0.1.16
#langchain-classic<1.0.0
langchain-community==0.0.34
langchain-core==0.1.53
langchain-openai==0.3.19
langchain-text-splitters==0.0.2
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.1.147
lightgbm==4.6.0
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==2.4.0
nest-asyncio==1.6.0
oauthlib==3.3.1
openai==2.6.1
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.10.18
ormsgpack==1.11.0
packaging==23.2
pandas==2.3.2
parso==0.8.4
pathspec==0.12.1
#pdfminer==20191125
pdfminer.six==20201018
pendulum==3.1.0
pg8000==1.31.5
pillow==11.3.0
platformdirs==4.3.8
prefect==3.4.17
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyodbc==5.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
python-socks==2.7.2
pytz==2025.2
#pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.1
scramp==1.4.6
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
streamlit==1.49.1
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uv==0.8.17
uvicorn==0.35.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
xgboost==3.0.5
xxhash==3.6.0
yarl==1.20.0
zipp==3.23.0
zstandard==0.23.0
chardet==5.2.0
aiofiles==23.2.1
